Fix integer→float division artifacts and enforce SpotBugs ICAST_IDIV_CAST_TO_DOUBLE
### Motivation
- Prevent incorrect numeric results caused by integer division being cast to `float`/`double` in rendering and chart math.
- Silence SpotBugs `ICAST_IDIV_CAST_TO_DOUBLE` findings by fixing actual occurrences rather than masking them.
- Preserve the intended rendering/geometry semantics while ensuring explicit floating-point arithmetic where required.
- Improve precision in pinch-zoom, layout centering, camera transforms, and UI metrics.

### Description
- Replace integer division with explicit floating-point operations in chart and layout code such as `ChartComponent.pointerDragged`, `XYChart` center/translate, and `RadarChart` center calculations to avoid truncation.
- Adjust UI/graphics math in `DefaultLookAndFeel`, `RoundRectBorder`, `CodenameOneImplementation.fillRectRadialGradient`, `FlipTransition`, and `PerspectiveCamera` to use `0.5f`/`... * 0.5f` or explicit casts where appropriate.
- Fix numeric coefficient calculation by casting factorial division to `double` in `Geometry` and correct `StyleParser.getMMValue` to return a properly computed `float`; also refine `CommonProgressAnimations.CircleProgress` timing calculation.
- Add `ICAST_IDIV_CAST_TO_DOUBLE` to the forbidden SpotBugs rules in `.github/scripts/generate-quality-report.py` so such findings will fail the quality gate.

### Testing
- No unit or integration test suite was executed as part of this change.
- Static SpotBugs analysis was the motivating CI check and the edits target the previously reported `ICAST_IDIV_CAST_TO_DOUBLE` violations.
- The repository previously failed the SpotBugs quality gate due to these violations and the changes remove the identified integer→float division instances.
- CI re-run of the full pipeline was not performed within this patch set.

+++ CodenameOne/src/com/codename1/charts/ChartComponent.java
@@ -571,19 +571,19 @@ public class ChartComponent extends Component {
                 }
             } else {
                 if (zoomStart == null) {
-                    zoomStart = new Point((x[0] + x[1]) / 2, (y[0] + y[1]) / 2);
+                    zoomStart = new Point((x[0] + x[1]) / 2f, (y[0] + y[1]) / 2f);
                     zoomTransformStart = Transform.makeIdentity();
                     if (transform != null) {
                         zoomTransformStart.concatenate(transform);
                     }
-                    int dx = Math.abs(x[0] - x[1]) / 2;
-                    int dy = Math.abs(y[0] - y[1]) / 2;
+                    double dx = Math.abs(x[0] - x[1]) / 2f;
+                    double dy = Math.abs(y[0] - y[1]) / 2f;
                     zoomDistStart = Math.sqrt(dx * dx + dy * dy);
 
 
                 } else {
-                    int dx = Math.abs(x[0] - x[1]) / 2;
-                    int dy = Math.abs(y[0] - y[1]) / 2;
+                    double dx = Math.abs(x[0] - x[1]) / 2f;
+                    double dy = Math.abs(y[0] - y[1]) / 2f;
                     double zoomDist = Math.sqrt(dx * dx + dy * dy);
                     if (zoomDist == 0) {
                         zoomDist = 1;
